requirements.txt
python-dotenv
feedparser
requests
google-generativeai
edge-tts
Pillow
google-cloud-texttospeech
moviepy
numpy
playwright
google-auth-oauthlib
google-api-python-client